backend/requirements.txt
# Core Django
Django==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS headers
django-cors-headers==4.3.1

# Image handling
Pillow==10.2.0

# Filtering
django-filter==23.5

# Async tasks
celery==5.3.6
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
django-debug-toolbar==4.2.0
